Make use of open_file from #84 Refactor done in #80 essentially reverted changes Re-fixes #81, also re-fixes windows opening behind TagStudio
