+++ tagstudio/src/qt/helpers/__init__.py
@@ -1,4 +1,3 @@
-from .open_file import open_file
-from .file_opener import FileOpenerHelper, FileOpenerLabel
+from .file_opener import open_file, FileOpenerHelper, FileOpenerLabel
 from .function_iterator import FunctionIterator
 from .custom_runnable import CustomRunnable
